+++ src/pages/ApplicationForm.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import { PageTemplate, MobileSizeContainer } from "../styles/template";
-import Form from "../components/Form";
+import Form from "../components/form/Form";
 
 type Props = {};
